Fix : Prevent default project from being deleted

+++ backend/apps/kloudust/lib/cmd/deleteProject.js
@@ -18,8 +18,8 @@ const CMD_CONSTANTS = require(`${KLOUD_CONSTANTS.LIBDIR}/cmd/cmdconstants.js`);
 module.exports.exec = async function(params) {
     if (!roleman.checkAccess(roleman.ACTIONS.edit_org)) { 
         params.consoleHandlers.LOGERROR("User is unauthorized for this operation."); return CMD_CONSTANTS.FALSE_RESULT(); }
-        
     const project = roleman.getNormalizedProject(params[0]||KLOUD_CONSTANTS.env.prj());
+    if(project === KLOUD_CONSTANTS.DEFAULT_PROJECT) { params.consoleHandlers.LOGWARN("Cannot delete project default."); return CMD_CONSTANTS.FALSE_RESULT("Cannot delete project default."); }
     const org = roleman.getNormalizedOrg(params[1]||KLOUD_CONSTANTS.env.org());
     return {result: await dbAbstractor.deleteProject(project, org), err: "", out: ""};
 }
